# docs/swagger.yaml
definitions:
  handler.Response:
    properties:
      code:
        type: integer
      data:
        type: object
      message:
        type: string
    type: object
  model.UserInfo:
    properties:
      createdOn:
        type: string
      email:
        type: string
      id:
        type: integer
      modifiedOn:
        type: string
      phone:
        type: string
      username:
        type: string
    type: object
  user.CreateRequest:
    properties:
      email:
        example: 82345611@gmail.com
        type: string
      password:
        example: "123456"
        type: string
      phone:
        example: "15845451515"
        type: string
      username:
        example: admin
        type: string
    type: object
  user.LoginRequest:
    properties:
      password:
        example: "123456"
        type: string
      username:
        example: admin
        type: string
    type: object
host: api.ywbang.top
info:
  contact: {}
  description: An example of gin Admin
  license:
    name: MIT
    url: https://github.com/liuvigongzuoshi/go-gin-backend-admin/blob/master/LICENSE
  termsOfService: https://github.com/liuvigongzuoshi/go-gin-backend-admin
  title: Golang Gin API
  version: "1.0"
paths:
  /user:
    get:
      parameters:
      - description: token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: '{"code":20000,"message":"OK","data": UserInfo }'
          schema:
            $ref: '#/definitions/model.UserInfo'
        "20003":
          description: '{"code":20003,"message":"Token has timed out.","data": null
            }'
          schema:
            $ref: '#/definitions/handler.Response'
      summary: 获取当前登录用户的用户信息
      tags:
      - 用户模块
  /user/login:
    post:
      parameters:
      - description: 登陆授权
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/user.LoginRequest'
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: '{"code":20000,"message":"OK","data":{"token":"eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpYXQiOjE1MjgwMTY5MjIsImlkIjowLCJuYmYiOjE1MjgwMTY5MjIsInVzZXJuYW1lIjoiYWRtaW4ifQ.LjxrK9DuAwAzUD8-9v43NzWBN7HXsSLfebw92DKd1JQ"}}'
          schema:
            $ref: '#/definitions/handler.Response'
      summary: 登陆授权
      tags:
      - 用户模块
  /user/register:
    post:
      consumes:
      - application/json
      parameters:
      - description: 用户注册
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/user.CreateRequest'
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: '{"code":20000,"data":{"name": "admin"},"msg":"ok"}'
          schema:
            $ref: '#/definitions/handler.Response'
      summary: 用户注册
      tags:
      - 用户模块
swagger: "2.0"